Prevent enemyattack() from being called after the attack is over
enemyattack() creates two background timers:
* The first waits 200ms and then actually damages the player.
  If the player's HP is reduced to zero, enemyattack() is called again stop the battle and close the window.

* The second waits a random period of time (2-3 seconds), then calls enemyattack().

The upshot is that when the player's health is reduced to zero by an attack, the window is immediately closed and the battle is stopped.
However, 2-3 (well, 1.8-2.8) seconds later, the second timer will fire and attempt to end the battle *again*.  Things go amiss if
a new battle has been started.

Solution: If the first timer shuts things down, have it abort the second timer.